rewrites estimator based gans
