+++ galaxy2galaxy/models/gans.py
@@ -1,10 +1,7 @@
-""" Spectral Norm GAN """
 
 from __future__ import absolute_import
 from __future__ import division
 from __future__ import print_function
-
-from functools import partial
 
 import tensorflow as tf
 import tensorflow.contrib.gan as tfgan
@@ -19,19 +16,14 @@ from tensorflow.python.ops import math_ops
 from tensorflow.python.ops.losses import losses
 from tensorflow.python.util import tf_inspect as inspect
 
-from tensor2tensor.layers import common_attention
-from tensor2tensor.layers import common_hparams
-from tensor2tensor.layers import common_layers
-from tensor2tensor.layers import discretization
-from tensor2tensor.layers import latent_layers
-from tensor2tensor.layers import modalities
 from tensor2tensor.utils import t2t_model
 from tensor2tensor.models import vanilla_gan
-from tensor2tensor.models.research.transformer_vae import residual_conv
 from tensor2tensor.layers.common_layers import lrelu
+from tensor2tensor.utils import hparams_lib
+from tensor2tensor.layers import common_layers
 
 from galaxy2galaxy.utils import registry
-from galaxy2galaxy.models.gan_utils import softplus_discriminator_loss, softplus_generator_loss, SperctraNormConstraint
+from galaxy2galaxy.models.gan_utils import softplus_discriminator_loss, softplus_generator_loss, SperctralNormConstraint
 
 
 def pack_images(images, rows, cols):
@@ -49,7 +41,6 @@ def pack_images(images, rows, cols):
     images = tf.transpose(images, [0, 2, 1, 3, 4])
     images = tf.reshape(images, [1, rows * width, cols * height, depth])
     return images
-
 
 @registry.register_model
 class SlicedGanLarge(vanilla_gan.SlicedGan):
@@ -162,64 +153,162 @@ class SlicedGanLarge(vanilla_gan.SlicedGan):
       net = tf.layers.conv2d(net, 3, (3,3), padding='SAME', name='output_conv')
 
       out = tf.nn.sigmoid(net)
-      return out #common_layers.convert_real_to_rgb(out)
+      return out
 
 @registry.register_model
-class SnGAN(SlicedGanLarge):
-  """Spectral Norm GAN"""
+class GanEstimator(SlicedGanLarge):
+    """ GAN based on tfgan estimator API
+    """
 
-  def body(self, features):
-    """Body of the model.
+  def discriminator(self, x, is_training, reuse=False,
+                    output_size=1024):
+    """Discriminator architecture with Spectral Normalization.
 
     Args:
-      features: a dictionary with the tensors.
+      x: input images, shape [bs, h, w, channels]
+      is_training: boolean, are we in train or eval model.
+      reuse: boolean, should params be re-used.
 
     Returns:
-      A pair (predictions, losses) where predictions is the generated image
-      and losses is a dictionary of losses (that get added for the final loss).
+      out_logit: the output logits (before sigmoid).
     """
-    features["targets"] = features["inputs"]
-    is_training = self.hparams.mode == tf.estimator.ModeKeys.TRAIN
+    hparams = self.hparams
+    with tf.variable_scope(
+        "discriminator", reuse=reuse,
+        initializer=tf.random_normal_initializer(stddev=0.02)):
+      batch_size, height, width = common_layers.shape_list(x)[:3]
+      do_update = is_training and (not reuse)
 
-    # Input images.
-    real_data = tf.to_float(features["targets_raw"])
-    out_shape = common_layers.shape_list(real_data)[1:4]
+      # Mapping x from [bs, h, w, c] to [bs, 1]
+      net = tf.layers.conv2d(x, 32, (3, 3), strides=(1, 1),
+                             padding="SAME", name="d_conv1",
+                             kernel_constraint=SperctralNormConstraint(update=do_update,
+                                                                       name='sn1'))
+      net = lrelu(net)
+      net = tf.layers.conv2d(net, 64, (4, 4), strides=(2, 2),
+                             padding="SAME", name="d_conv1b",
+                             kernel_constraint=SperctralNormConstraint(update=do_update,
+                                                                       name='sn1b'))
+      # [bs, h/2, w/2, 64]
+      net = lrelu(net)
+      net = tf.layers.conv2d(net, 64, (3, 3), strides=(1, 1),
+                             padding="SAME", name="d_conv2",
+                             kernel_constraint=SperctralNormConstraint(update=do_update,
+                                                                       name='sn2'))
+      net = lrelu(net)
+      net = tf.layers.conv2d(net, 128, (4, 4), strides=(2, 2),
+                             padding="SAME", name="d_conv2b",
+                             kernel_constraint=SperctralNormConstraint(update=do_update,
+                                                                       name='sn2b'))
+      # [bs, h/4, w/4, 128]
+      net = lrelu(net)
+      net = tf.layers.conv2d(net, 128, (3, 3), strides=(1, 1),
+                             padding="SAME", name="d_conv3",
+                             kernel_constraint=SperctralNormConstraint(update=do_update,
+                                                                       name='sn3'))
+      net = lrelu(net)
+      net = tf.layers.conv2d(net, 256, (4, 4), strides=(2, 2),
+                             padding="SAME", name="d_conv3b",
+                             kernel_constraint=SperctralNormConstraint(update=do_update,
+                                                                       name='sn3b'))
+      # [bs, h/8, w/8, 256]
+      net = lrelu(net)
+      net = tf.layers.flatten(net)
+      net = tf.layers.dense(net, output_size, name="d_fc3",
+                             kernel_constraint=SperctralNormConstraint(update=do_update,
+                                                                       name='sn4'))  # [bs, 1024]
 
-    # Noise vector.
+      net = lrelu(net)
+      return net
+
+  @classmethod
+  def estimator_model_fn(cls,
+                         hparams,
+                         features,
+                         labels,
+                         mode,
+                         config=None,
+                         params=None,
+                         decode_hparams=None,
+                         use_tpu=False):
+
+    if mode not in [model_fn_lib.ModeKeys.TRAIN, model_fn_lib.ModeKeys.EVAL,
+                  model_fn_lib.ModeKeys.PREDICT]:
+      raise ValueError('Mode not recognized: %s' % mode)
+
+    if mode is model_fn_lib.ModeKeys.TRAIN:
+      is_training = True
+      hparams = hparams_lib.copy_hparams(hparams)
+
+    # Instantiate model
+    data_parallelism = None
+    if not use_tpu and config:
+      data_parallelism = config.data_parallelism
+    reuse = tf.get_variable_scope().reuse
+
+    # Instantiate model
+    self = cls(
+        hparams,
+        mode,
+        data_parallelism=data_parallelism,
+        decode_hparams=decode_hparams,
+        _reuse=reuse)
+
+    real_data =  common_layers.convert_rgb_to_real(features['inputs'])  # rename inputs for clarity
     generator_inputs = tf.random_uniform([self.hparams.batch_size,
                                           self.hparams.bottleneck_bits],
                                           minval=-1, maxval=1, name="z")
 
-    # Manual gan_model creation
-    generated_images = vanilla_gan.reverse_gradient(self.generator(generator_inputs, is_training, out_shape))
+    # rename inputs for clarity
+    out_shape = common_layers.shape_list(real_data)[1:4]
 
-    tf.logging.info("Shape of generated images", generated_images.shape )
-    discriminator_gen_outputs = self.discriminator(common_layers.convert_rgb_to_symmetric_real(generated_images), is_training, output_size=1)
-    discriminator_real_outputs = self.discriminator(common_layers.convert_rgb_to_symmetric_real(real_data), is_training, reuse=True, output_size=1)
+    if mode == model_fn_lib.ModeKeys.PREDICT:
+      if real_data is not None:
+        raise ValueError('`labels` must be `None` when mode is `predict`. '
+                           'Instead, found %s' % real_data)
+      gan_model = _make_prediction_gan_model(generator_inputs,
+                                                   partial(self.generator, is_training=is_training, out_shape=out_shape),
+                                                   'Generator')
+    # Here should be where we export the model as tf hub
 
-    gan_model = tfgan.GANModel(
-        generator_inputs, generated_images,
-        None, None, None,
-        real_data, discriminator_real_outputs, discriminator_gen_outputs,
-        None, None, None)
+    else:  # model_fn_lib.ModeKeys.TRAIN or model_fn_lib.ModeKeys.EVAL
+            # Manual gan_model creation
+      with tf.variable_scope('Generator') as gen_scope:
+        generated_images = self.generator(generator_inputs, is_training=is_training, out_shape=out_shape)
 
-    losses = tfgan.gan_loss(gan_model,
-                    generator_loss_fn=softplus_generator_loss,
-                    discriminator_loss_fn=softplus_discriminator_loss)
+      with tf.variable_scope('Discriminator') as dis_scope:
+        discriminator_gen_outputs = self.discriminator(generated_images, is_training=is_training)
 
-    losses = {'training': losses.generator_loss + losses.discriminator_loss,
-              'gen_loss': losses.generator_loss,
-              'disc_loss': losses.discriminator_loss}
+      with tf.variable_scope(dis_scope, reuse=True):
+        discriminator_real_outputs = self.discriminator(real_data, is_training=is_training)
 
-    summary_g_image = tf.reshape(
-        generated_images[0, :], [1] + common_layers.shape_list(real_data)[1:])
+      tf.summary.image("generated", pack_images(generated_images, 4, 4), max_outputs=1)
+      tf.summary.image("real", pack_images(real_data, 4, 4), max_outputs=1)
 
-    tf.summary.image("generated", pack_images(generated_images, 4, 4), max_outputs=1)
-    tf.summary.image("real", pack_images(real_data, 4, 4), max_outputs=1)
+      generator_variables = variable_lib.get_trainable_variables(gen_scope)
+      discriminator_variables = variable_lib.get_trainable_variables(dis_scope)
 
-    if is_training:  # Returns an dummy output and the losses dictionary.
-      return tf.zeros_like(real_data), losses
-    return tf.reshape(generated_images, tf.shape(real_data)), losses
+      gan_model = tfgan.GANModel(
+                generator_inputs,
+                generated_images,
+                generator_variables,
+                gen_scope,
+                self.generator,
+                real_data,
+                discriminator_real_outputs,
+                discriminator_gen_outputs,
+                discriminator_variables,
+                dis_scope,
+                self.discriminator)
+
+      opt_gen = tf.train.AdamOptimizer(hparams.learning_rate)
+      opt_disc = tf.train.AdamOptimizer(hparams.learning_rate)
+
+    # Make the EstimatorSpec, which incorporates the GANModel, losses, eval
+    # metrics, and optimizers (if required).
+    return _get_estimator_spec(
+      mode, gan_model, softplus_generator_loss, softplus_discriminator_loss,
+      None, opt_gen, opt_disc, None, True)
 
 @registry.register_hparams
 def gan_large():
